fix: remove illegal 'as any' casts and split AllProviders into its own file
- Metrics.tsx: cast metricValue to Json type instead of using 'as any'
  on the entire insert object (fixes CI error on line R249)
- Profile.tsx: remove 'as any' from .eq() and .maybeSingle() calls and
  from .upsert() — Supabase types already accept the correct column
  names and data shapes (fixes CI errors on lines R43/R44/R165)
- History.tsx: remove 'as any' from .eq('id', ...) in both toggleResolved
  and deleteEntry (fixes CI errors on lines R136/R165)
- src/test/AllProviders.tsx: extract AllProviders component into its own
  file so utils.tsx only contains non-component exports, which resolves
  the fast-refresh ESLint warning (R60 / R90)

Fixes #274
